+++ src/types/pre_checkout_query.rs
@@ -1,4 +1,5 @@
 use crate::types::order_info::OrderInfo;
+use crate::types::primitive::Integer;
 use crate::types::user::User;
 
 /// This object contains information about an incoming pre-checkout query.
@@ -17,7 +18,7 @@ pub struct PreCheckoutQuery {
     /// it shows the number of digits past the
     /// decimal point for each currency
     /// (2 for the majority of currencies).
-    pub total_amount: i64,
+    pub total_amount: Integer,
     /// Bot specified invoice payload
     pub invoice_payload: String,
     /// Identifier of the shipping option chosen by the user
