+++ src/types/video_note.rs
@@ -1,4 +1,5 @@
 use crate::types::photo_size::PhotoSize;
+use crate::types::primitive::Integer;
 
 /// This object represents a video message
 #[derive(Debug)]
@@ -6,11 +7,11 @@ pub struct VideoNote {
     /// Unique identifier for this file
     pub file_id: String,
     /// Video width and height
-    pub length: i64,
+    pub length: Integer,
     ///  Duration of the video in seconds
-    pub duration: i64,
+    pub duration: Integer,
     /// Video thumbnail
     pub thumb: Option<PhotoSize>,
     /// File size
-    pub file_size: Option<i64>,
+    pub file_size: Option<Integer>,
 }
